Add files via upload
Determining the relation between higher attributes of images, and the number of likes received by posts usinf them on social media platform Facebook. The attributes of fine object recognition,semantic segmentation and similarity to the image with the highest number of likes were determined for images extracted using Apify. Similarity is the attribute with the maximum amount of influence on the engagement gained by social media posts.